Made changes to how the User Interface runs to avoid redundant user interface updates.
User Interface updates will now occur twice every second, consistently as long as the application is in focus.

No updates will occur when the application looses focus (after onPause()).

Calls to the update function will be made even when the service is not available, but no actual updates will be made.