Among the changes made:
To avoid much memory allocation / garbage collection, the following have been changed:
1) The RecorderService now runs a thread that keeps sampling
	from being affected by UI activities.

2) The ClassifierService has been changed to ClassifierThread.
	This means the computations in the ClassifierThread
	don't have to affect the sampling in RecorderThread.
	RecorderThread can now sample down to 50ms + ~2ms error.
	Also garbage collection is invoked before the samples
	are taken to avoid sampling being interrupted by
	garbage collection.

3) The accelerometer reader
	Instead of
		float[] getSample()
	we have
		void assignSample(float[])
	The caller of the function provides the array to which
	the accelerometer reader should fill the values.
	Internally, the RealAccelReader uses 3 arrays, among which
	it iterates so as to avoid reading and writing to the
	same array at any one particular instance. This is
	of course enforced by using mutexes.
	This means that the accelerometer reader could be
	simultaniously updating one array while at the same
	time another thread could be reading the other array.

4) Introduction of SampleBatch and SampleBatchBuffer

The SampleBatch class holds 128 samples (3d vectors). Plus
any other required data or data gathered during analysis of the batch.
The SampleBatch is reusable, meaning that there is no need to create new
SampleBatch instances. Instances of SampleBatch are in use in the Sampler class,
RecorderService, and ClassifierThread. While not in use in any of these,
SampleBatch instances are placed in the SampleBatchBuffer.
The SampleBatchBuffer also serves as a two-way blocking pipe between
the RecorderService and the ClassifierThread. The RecorderService
obtains empty SampleBatch instances (without any data) and fills them with
sampled data, while the ClassifierThread gets the filled
SampleBatch instances and analyses them to come up with a classification.
The ClassifierThread blocks until a filled SampleBatch instance is
available.

5) Change of sampled data from float[384] to float[128][3].